requirements.txt
azure-ai-documentintelligence>=1.0.2
azure-identity>=1.17.1
